api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
qdrant-client==1.11.0
sentence-transformers==3.1.1
psycopg2-binary==2.9.9
redis==5.0.8
celery==5.4.0
httpx==0.27.2
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
langgraph==0.2.55
langchain-core==0.3.25
langfuse==2.55.0
